+++ lib/Router.class.ts
@@ -3,6 +3,7 @@ import {Middleware} from "./Middleware.class";
 import {Application} from "./Application.class";
 import * as url from 'url';
 import {Context} from "./Context.class";
+import {Stream} from "stream";
 
 export namespace Router {
     export class Router {
@@ -10,25 +11,9 @@ export namespace Router {
         protected params = {};
 
         constructor(protected options: IOptions) {
-            if (options.httpHandler === true) {
-                this.use(async (ctx, next) => {
-                    try {
-                        await next();
-                    } catch (err) {
-                        ctx.json({
-                            status: 500,
-                            error: {
-                                name: err.name,
-                                message: err.message,
-                                stack: err.stack
-                            },
-                        });
-                    }
-                });
-            }
         }
 
-        use(path: Path | Middleware.Middleware | Middleware.Middleware[] | null, middleware?: Middleware.Middleware | Middleware.Middleware[]) {
+        public use(path: Path | Middleware.Middleware | Middleware.Middleware[] | null, middleware?: Middleware.Middleware | Middleware.Middleware[]) {
             middleware = typeof path === 'function' ? path : middleware;
             const hasPath = typeof path === 'string';
 
@@ -67,13 +52,16 @@ export namespace Router {
                         setRouterParams(Object.keys(this.params));
                     }
                 } else {
-                    this.register(hasPath ? path as Path : '(.*)', mw, {end: false, ignoreCaptures: !hasPath} as IOptions);
+                    this.register(hasPath ? path as Path : '(.*)', mw, {
+                        end: false,
+                        ignoreCaptures: !hasPath
+                    } as IOptions);
                 }
             });
             return this;
         }
 
-        register(path: Path, middleware: Middleware.Middleware | Middleware.Middleware[], options: IOptions = {} as IOptions) {
+        public register(path: Path, middleware: Middleware.Middleware | Middleware.Middleware[], options: IOptions = {} as IOptions) {
             const stack = this.stack;
 
             const route = new Route.Route(path, middleware, {
@@ -100,7 +88,7 @@ export namespace Router {
             return route;
         }
 
-        route(name) {
+        public route(name) {
             const routes = this.stack;
             for (let len = routes.length, i = 0; i < len; i++) {
                 if (routes[i].name && routes[i].name === name) return routes[i];
@@ -108,7 +96,7 @@ export namespace Router {
             return false;
         }
 
-        match(path) {
+        public match(path) {
             const routes = this.stack;
             let route: Route.Route;
             const matched = {
@@ -128,7 +116,7 @@ export namespace Router {
             return matched;
         }
 
-        param(param, middleware) {
+        public param(param, middleware) {
             this.params[param] = middleware;
             for (let i = 0; i < this.stack.length; i++) {
                 const route = this.stack[i];
@@ -138,13 +126,16 @@ export namespace Router {
             return this;
         }
 
-        routes() {
+        public routes(http: boolean = false) {
             const router = this;
 
             let dispatch: Middleware.Middleware = (ctx, next) => {
                 const method = ctx.method.toLowerCase();
 
-                if(method && router.options.httpHandler) router.use(ctx => { ctx.res.status = 404; ctx.json({ json: 404 })  });
+                if (method && http) router.use(ctx => {
+                    ctx.res.status = 404;
+                    ctx.json({json: 404})
+                });
 
                 const path = router.options.routerPath || ctx.routerPath || ctx.path;
                 const matched = router.match(path);
@@ -175,13 +166,15 @@ export namespace Router {
                         return next();
                     });
                     return memo.concat(
-                        method && router.options.httpHandler ?
+                        method && http ?
                             route.stack.filter(mw => mw.method === method || !mw.method) :
                             route.stack
                     );
                 }, [] as Middleware.Middleware[]);
 
-                return Application.Application.createCompose(routeChain)(ctx, next);
+                const composed = Application.Application.createCompose(routeChain)(ctx, next);
+
+                return http ? composed.then(() => ctx.__handleEnd()).catch(ctx.error) : composed;
             };
 
             dispatch.router = router;
@@ -189,80 +182,7 @@ export namespace Router {
             return dispatch;
         }
 
-        // Simple http routes handling.
-        httpRoutes() {
-            if(!this.options.httpHandler) throw new Error('Router should be with option `httpHandler: true`');
-            return (request, response) => {
-                const ctx = Object.create(new Context.Context({ path: url.parse(request.url).pathname, app: { options: {} } as Application.Application }));
-
-                const req = {
-                    req: request,
-                    res: response,
-                    get method() { return this.req.method; },
-                    get url() { return this.req.url; },
-                    get path() { return url.parse(this.url).pathname; },
-                };
-
-                const res = {
-                    ctx,
-                    req: request,
-                    res: response,
-                    app: ctx.app,
-                    get headersSent() {
-                        return this.res.headersSent;
-                    },
-                    get status() {
-                        return this.res.statusCode;
-                    },
-                    set status(code) {
-                        ctx.assert(
-                            code >= 100 && code <= 999 &&
-                            Number.isInteger(code) &&
-                            Number.isFinite(code),
-                            new Error(`Invalid status code: ${code}, must be a number & not out of range 100 ~ 999.`)
-                        );
-                        this.res.statusCode = code;
-                    },
-                    get message() {
-                        return this.res.statusMessage;
-                    },
-                    set message(message) {
-                        this.res.statusMessage = message;
-                    },
-                    send(...args) {
-                        this.end(...args);
-                    },
-                    end(...args) {
-                        this.res.end(...args);
-                    },
-                    json(obj) {
-                        this.status = 200;
-                        this.res.setHeader('Content-Type', 'application/json');
-                        this.res.end(JSON.stringify(obj));
-                    },
-                    redirect(url) {
-                        this.res.setHeader('Location', url);
-                        this.res.setHeader('Content-Type', 'text/plain; charset=utf-8');
-                        this.status = 302;
-                        this.end('Redirecting...');
-                    }
-                };
-
-                ctx.request = request;
-                ctx.req = req;
-                ctx.response = response;
-                ctx.res = res;
-                ctx.url = req.url;
-                ctx.method = req.method;
-                ctx.send = res.send.bind(res);
-                ctx.end = res.end.bind(res);
-                ctx.json = res.json.bind(res);
-                ctx.redirect = res.redirect.bind(res);
-                this.routes()(ctx, (err) => { ctx.next && ctx.next(err); });
-            };
-        }
-
-        process(name?: any, path?: any, middleware?: Middleware.Middleware | Middleware.Middleware[]) {
+        public process(name?: any, path?: any, middleware?: Middleware.Middleware | Middleware.Middleware[]) {
             if (typeof path === "string" || path instanceof RegExp) {
                 middleware = Array.prototype.slice.call(arguments, 2);
             } else {
@@ -279,6 +199,272 @@ export namespace Router {
         }
     }
 
+    export class HttpRouter extends Router {
+        constructor(options: IOptions) {
+            super(options);
+            this.use(async (ctx, next) => {
+                try {
+                    await next();
+                } catch (err) {
+                    ctx.json({
+                        status: 500,
+                        error: {
+                            name: err.name,
+                            message: err.message,
+                            stack: err.stack
+                        },
+                    });
+                }
+            });
+        }
+
+        protected extendContext(ctx, request, response) {
+            Object.defineProperties(ctx, {
+                originRequest: {
+                    value: request,
+                    configurable: false
+                },
+                originResponse: {
+                    value: response,
+                    configurable: false
+                },
+                req: {
+                    value: request,
+                    configurable: false
+                },
+                res: {
+                    value: response,
+                    configurable: false
+                },
+                request: {
+                    value: request,
+                    configurable: false
+                },
+                response: {
+                    value: response,
+                    configurable: false
+                },
+                method: {
+                    get() {
+                        return request.method;
+                    },
+                    configurable: false,
+                },
+                url: {
+                    get() {
+                        return request.url;
+                    },
+                    configurable: false,
+                },
+                path: {
+                    get() {
+                        return url.parse(request.url).pathname;
+                    },
+                    configurable: false,
+                },
+                headerSent: {
+                    get() {
+                        return response.headersSent;
+                    },
+                    configurable: false,
+                },
+                status: {
+                    get() {
+                        return response.statusCode;
+                    },
+                    set(code) {
+                        ctx.assert(
+                            code >= 100 && code <= 999 &&
+                            Number.isInteger(code) &&
+                            Number.isFinite(code),
+                            new Error(`Invalid status code: ${code}, must be a number & in range 100 ~ 999.`)
+                        );
+                        response.statusCode = code;
+                    },
+                    configurable: false,
+                },
+                message: {
+                    get() {
+                        return response.statusMessage;
+                    },
+                    configurable: false,
+                },
+                body: {
+                    get() {
+                        return this._body;
+                    },
+                    set(value) {
+                        const body = this._body = value;
+
+                        if (value === null) {
+                            this.status = 204;
+                            this.remove('Content-Type');
+                            this.remove('Content-Length');
+                            this.remove('Transfer-Encoding');
+                            return;
+                        }
+
+                        const hasContentType = !this.has('Content-Type');
+
+
+                        if (Buffer.isBuffer(value)) {
+                            if (hasContentType) this.type = 'bin';
+                            this.length = value.length;
+                            return;
+                        }
+
+                        if (value instanceof Stream) {
+                            response.socket.once('finish', () => {
+                                this.__responseEnded = true;
+                                response.destroy();
+                            });
+                            if (body != value) {
+                                value.once('error', err => this.ctx.onerror(err));
+                                // overwriting
+                                if (body !== null) this.remove('Content-Length');
+                            }
+
+                            if (hasContentType) this.type = 'bin';
+                            return;
+                        }
+
+                        if (typeof value === 'string') {
+                            if (hasContentType) this.type = /^\s*</.test(value) ? 'html' : 'text';
+                            this.length = Buffer.byteLength(value);
+                            return;
+                        }
+                        this.remove('Content-Length');
+                        this.type = 'json';
+                    },
+                    configurable: false,
+                },
+                remove: {
+                    value: function (field) {
+                        if (this.headerSent) return;
+                        response.removeHeader(field);
+                    },
+                    configurable: false
+                },
+                set: {
+                    value: function set(field, val) {
+                        if (this.headerSent) return;
+
+                        if (arguments.length === 2) {
+                            if (Array.isArray(val)) val = val.map(v => typeof v === 'string' ? v : String(v));
+                            else if (typeof val !== 'string') val = String(val);
+                            response.setHeader(field, val);
+                        } else {
+                            for (const key in field) {
+                                this.set(key, field[key]);
+                            }
+                        }
+                    },
+                    configurable: false
+                },
+                get: {
+                    value: function (field) {
+                        const headers = typeof response.getHeaders === 'function'
+                            ? response.getHeaders()
+                            : response._headers || {};
+
+                        if (!field) return headers;
+                        return headers[field.toLowerCase()] || '';
+                    },
+                    configurable: false
+                },
+                has: {
+                    value: function (field) {
+                        return typeof response.hasHeader === 'function'
+                            ? response.hasHeader(field)
+                            : field.toLowerCase() in this.headers;
+                    },
+                    configurable: false
+                },
+                type: {
+                    get() {
+                        const contentType = this.get('Content-Type');
+                        if (!contentType) return '';
+                        return contentType.split(';', 1)[0];
+                    },
+                    set(contentType) {
+                        if (contentType) {
+                            this.set('Content-Type', contentType);
+                        } else {
+                            this.remove('Content-Type');
+                        }
+                    },
+                    configurable: false,
+                },
+                length: {
+                    set(len) {
+                        this.set('Content-Length', len);
+                    },
+                    configurable: false,
+                },
+                send: {
+                    value(data) {
+                        response.end(data);
+                    },
+                    configurable: false
+                },
+                end: {
+                    value(data) {
+                        response.end(data);
+                    },
+                    configurable: false
+                },
+                json: {
+                    value(data) {
+                        response.end(JSON.stringify(data));
+                    },
+                    configurable: false
+                },
+                __responseEnded: {
+                    set(val: boolean) {
+                        this._responseEnded = val;
+                    },
+                    get() {
+                        return this._responseEnded;
+                    },
+                    configurable: false,
+                    enumerable: false
+                },
+                __handleEnd: {
+                    value() {
+                        if (!this._responseEnded) {
+                            if (this._body instanceof Stream) return this._body.pipe(response);
+                            if (Buffer.isBuffer(this._body)) return response.end(this._body);
+                            if (typeof this._body === 'string') return response.end(this._body);
+
+                            this.body = JSON.stringify(this._body);
+                            if (!response.headersSent) {
+                                ctx.length = Buffer.byteLength(this._body);
+                            }
+                            response.end(this._body);
+                        }
+                    },
+                    enumerable: false,
+                    configurable: false
+                }
+            });
+            return ctx;
+        }
+
+        // Simple http routes handling.
+        public routes() {
+            return (request, response) => {
+                const ctx = Object.create(new Context.Context({
+                    path: url.parse(request.url).pathname,
+                    app: {options: {}} as Application.Application
+                }));
+                this.extendContext(ctx, request, response);
+                super.routes(true)(ctx, (err) => {
+                    ctx.next && ctx.next(err);
+                });
+            };
+        }
+    }
+
     export interface IOptions {
         prefix?: string;
         strict?: string;
@@ -287,7 +473,6 @@ export namespace Router {
         end?: boolean;
         name?: string;
         routerPath?: string;
-        httpHandler?: boolean;
     }
 
     export const PathSeparator = '/';